[NRR-6536] Rename Nylas v2 lib path
