+++ .bumpversion.cfg
@@ -1,6 +1,6 @@
 [bumpversion]
 commit = True
 tag = True
-current_version = 5.14.1
+current_version = 5.14.2
 
-[bumpversion:file:nylas/_client_sdk_version.py]
+[bumpversion:file:nylasv2/_client_sdk_version.py]
