+++ .coveragerc
@@ -1,2 +1,2 @@
 [run]
-source = nylas
+source = nylasv2
