--- nylas/_client_sdk_version.py
@@ -1,1 +1,0 @@
-__VERSION__ = "5.14.1"
