+++ nylasv2/_client_sdk_version.py
@@ -1,0 +1,1 @@
+__VERSION__ = "5.14.2"
